Fleshed out Readme.md
Improved the Mono Warning, expanded it to .Net Core, and moved it closer to the top of the Read Me.
Fixed example code
Adjusted todo list, and removed 'Finding a way to deal with Mono's security issue', as because this is not a rabbit hole i want to go down, and frankly due to the way Vndb accepts logins, any implementation will be flawed because it has to be exposed as a plain text string at some point or another. It is just there to reduce attack vectors.