Update VndbFilters.cs - Raw-Filter
Added the Raw filter which allows providing the raw name and value, is always valid as far as VndbSharp is concerned, which means it doesn't try to validate the operator for the filter.

The reason this filter was added, was while fixing #42, i realized that it's not possible for the filters i've made to handle every single scenario so providing the end user the ability to resort to manually providing filters is the next best option. Of course, i could have just left it out, since the `AbstractFilter<T>` class does all the heavy lifting of creating a filter.... But why not?